requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
PyPDF2
tiktoken